refactor(src): change jsx files to js
