+++ src/common/Form/Input/index.js
@@ -1,1 +1,1 @@
-export { default } from './Input.jsx';
+export { default } from './Input.js';
