+++ src/common/Home/index.js
@@ -1,1 +1,1 @@
-export { default } from './Home.jsx';
+export { default } from './Home.js';
