+++ src/common/Nav/index.js
@@ -1,1 +1,1 @@
-export { default } from './Nav.jsx';
+export { default } from './Nav.js';
