+++ src/common/index.js
@@ -1,3 +1,3 @@
-export { default as createApp } from './create-app.jsx';
+export { default as createApp } from './create-app.js';
 
 export const ssrStateKey = 'SSRState';
